chore(html): refactor listbox spec and templates
